"DirectiveLiteral" as Literal
